Bug 1919688: Expand Metric enum to uint64_t from uint32_t r=denispal
Differential Revision: https://phabricator.services.mozilla.com/D237302

+++ tools/performance/PerfStats.h
@@ -77,7 +77,7 @@ class PerfStats {
  public:
   typedef MozPromise<nsCString, bool, true> PerfStatsPromise;
 
-  enum class Metric : uint32_t {
+  enum class Metric : uint64_t {
 #define DECLARE_ENUM(metric) metric,
     FOR_EACH_PERFSTATS_METRIC(DECLARE_ENUM)
 #undef DECLARE_ENUM
@@ -157,13 +157,14 @@ class PerfStats {
   static MetricMask sCollectionMask;
   static StaticMutex sMutex MOZ_UNANNOTATED;
   static StaticAutoPtr<PerfStats> sSingleton;
-  TimeStamp mRecordedStarts[static_cast<size_t>(Metric::Max)];
-  double mRecordedTimes[static_cast<size_t>(Metric::Max)];
-  uint32_t mRecordedCounts[static_cast<size_t>(Metric::Max)];
+  TimeStamp mRecordedStarts[static_cast<uint64_t>(Metric::Max)];
+  double mRecordedTimes[static_cast<uint64_t>(Metric::Max)];
+  uint32_t mRecordedCounts[static_cast<uint64_t>(Metric::Max)];
   nsTArray<nsCString> mStoredPerfStats;
 };
 
-static_assert(1 << (static_cast<uint64_t>(PerfStats::Metric::Max) - 1) <=
+static_assert(static_cast<uint64_t>(1)
+                      << (static_cast<uint64_t>(PerfStats::Metric::Max) - 1) <=
                   std::numeric_limits<PerfStats::MetricMask>::max(),
               "More metrics than can fit into sCollectionMask bitmask");
